python-backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn>=0.35,<1.0
pydantic>=2.11,<3.0
Pillow>=11.3,<12.0
httpx>=0.28,<1.0
PyInstaller>=6.16,<7.0
